Commit 24
Added more images to the program and fiddled with seconds and finalTime to get it right.  This will be my last code commit of the project and Overall I feel happy about this piece, but of course there are some tiny bugs still that I couldn't fix in time, but that's life for yah.